MODLD-746: Separate service for PreferredProfile
