+++ src/main/java/org/folio/linked/data/service/profile/ProfileService.java
@@ -1,8 +1,6 @@
 package org.folio.linked.data.service.profile;
 
 import java.util.List;
-import java.util.UUID;
-import javax.annotation.Nullable;
 import org.folio.linked.data.domain.dto.ProfileMetadata;
 
 public interface ProfileService {
@@ -13,8 +11,4 @@ public interface ProfileService {
   String getProfileById(Integer id);
 
   List<ProfileMetadata> getMetadataByResourceType(String resourceTypeUri);
-
-  void setPreferredProfile(UUID userId, Integer profileId, String resourceTypeUri);
-
-  List<ProfileMetadata> getPreferredProfiles(UUID userId, @Nullable String resourceTypeUri);
 }
